+++ src/sentry/api/validators/__init__.py
@@ -1,9 +1,6 @@
 from .auth import *  # noqa: F401,F403
 from .broadcast import *  # noqa: F401,F403
-from .doc_integration import *  # noqa: F401,F403
 from .email import *  # noqa: F401,F403
-from .external_actor import *  # noqa: F401,F403
-from .integrations import *  # noqa: F401,F403
 from .notifications import *  # noqa: F401,F403
 from .project import *  # noqa: F401,F403
 from .project_codeowners import *  # noqa: F401,F403
