+++ src/sentry/integrations/api/bases/doc_integrations.py
@@ -7,10 +7,10 @@ from rest_framework.request import Request
 
 from sentry.api.base import Endpoint
 from sentry.api.permissions import SentryPermission, StaffPermissionMixin
-from sentry.api.validators.doc_integration import METADATA_PROPERTIES
 from sentry.auth.superuser import is_active_superuser
 from sentry.integrations.api.bases.integration import PARANOID_GET
 from sentry.integrations.models.doc_integration import DocIntegration
+from sentry.integrations.validators.doc_integration import METADATA_PROPERTIES
 from sentry.utils.sdk import Scope
