+++ src/sentry/integrations/api/serializers/rest_framework/doc_integration.py
@@ -8,13 +8,13 @@ from rest_framework.serializers import Serializer, ValidationError
 
 from sentry.api.fields.avatar import AvatarField
 from sentry.api.helpers.slugs import sentry_slugify
-from sentry.api.validators.doc_integration import validate_metadata_schema
 from sentry.integrations.models.doc_integration import DocIntegration
 from sentry.integrations.models.integration_feature import (
     Feature,
     IntegrationFeature,
     IntegrationTypes,
 )
+from sentry.integrations.validators.doc_integration import validate_metadata_schema
 from sentry.sentry_apps.api.parsers.sentry_app import URLField
