+++ src/sentry/integrations/validators/external_actor.py
@@ -3,9 +3,9 @@ import re
 from rest_framework import serializers
 
 from sentry.api.exceptions import ParameterValidationError
-from sentry.api.validators.integrations import validate_provider
 from sentry.integrations.services.integration import integration_service
 from sentry.integrations.types import ExternalProviders
+from sentry.integrations.validators.integrations import validate_provider
 from sentry.models.organization import Organization
 
 EXTERNAL_ID_LENGTH_MIN = 1
